First step succeeded
Parse working for some cases like:
http://www.1999.co.jp/eng/10310668
http://www.1999.co.jp/eng/10024859

Needs to treat some cases like the MG series that the ‘Scale’ is not
just ‘1/100’, but ‘Master Grade, 1/100’. (e.g.:
http://www.1999.co.jp/eng/10055446)

Parser tries to pull the information from the website till it gets it.

Next steps: fix the Master Grade issues and iterate for the first 50
itens of the website.